增加移动端选项，并且增加前端的npm package

+++ generator-developer/generators/front/index.js
@@ -4,13 +4,31 @@ var Generator = require('yeoman-generator');
 module.exports = class extends Generator {
   constructor(args, opts) {
     super(args, opts);
+    this.props = {};
+  }
+
+  prompting() {
+    return this.prompt([{
+      type: 'confirm',
+      name: 'hybrid',
+      message: '是否为移动端项目?'
+    }]).then((answer) => {
+      this.props['hybrid'] = answer.hybrid;
+    })
   }
 
   configuring() {
-  	this.fs.copy(
-  		this.templatePath(),
-      this.destinationPath()
-  	)
+    if (this.props.hybrid) {
+      this.fs.copy(
+        this.templatePath('phone'),
+        this.destinationPath('./front')
+      )
+    } else {
+      this.fs.copy(
+        this.templatePath('web'),
+        this.destinationPath('./front')
+      )  
+    }
   }
 
   install() {
